requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
Brotli==1.0.9
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
constantly==15.1.0
cryptography==36.0.2
cssselect==1.1.0
dash==2.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-split-pane==1.0.0
dash-table==5.0.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dispatcher==1.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.6.0
Flask==2.1.2
Flask-Compress==1.12
huggingface-hub==0.4.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
incremental==21.3.0
ipykernel==6.12.1
ipython==8.2.0
ipython-genutils==0.2.0
itemadapter==0.5.0
itemloaders==1.0.4
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jmespath==1.0.0
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.2.1
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
langdetect==1.0.9
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
mysql-connector-python==8.0.28
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parsel==1.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
plotly==5.7.0
prometheus-client==0.14.0
prompt-toolkit==3.0.29
Protego==0.2.1
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
Pygments==2.11.2
pymongo==4.1.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
queuelib==1.6.2
regex==2022.3.15
requests==2.27.1
requests-file==1.5.1
sacremoses==0.0.49
Scrapy==2.6.1
scrapy-rotating-proxies==0.6.2
scrapy-splash==0.8.0
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
stem==1.8.0
tenacity==8.0.1
terminado==0.13.3
testpath==0.6.0
tldextract==3.2.0
tokenizers==0.11.6
torch==1.11.0
tornado==6.1
tqdm==4.63.1
traitlets==5.1.1
transformers==4.17.0
Twisted==22.2.0
typing==3.7.4.3
typing_extensions==4.1.1
urllib3==1.26.9
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
zipp==3.8.0
zope.interface==5.4.0
